+++ src/main/java/com/truelayer/java/HostedPaymentPageLinkBuilder.java
@@ -64,7 +64,7 @@ public class HostedPaymentPageLinkBuilder {
 
         URI hppLink = URI.create(MessageFormat.format(
                 "{0}/{1}#{2}={3}&resource_token={4}&return_uri={5}",
-                environment.getHppUri(),
+                getHppLinkForResourceType(resourceType, environment),
                 resourceType.getHppLinkPath(),
                 resourceType.getHppLinkQueryParameter(),
                 resourceId,
@@ -80,5 +80,19 @@ public class HostedPaymentPageLinkBuilder {
         }
 
         return hppLink;
+    }
+
+    private URI getHppLinkForResourceType(ResourceType resourceType, Environment environment) {
+        URI hostedPageUri;
+        switch (resourceType.getHostedPageType()) {
+            case HP2:
+                hostedPageUri = environment.getHp2Uri();
+                break;
+            case HPP:
+            default:
+                hostedPageUri = environment.getHppUri();
+                break;
+        }
+        return hostedPageUri;
     }
 }