+++ src/main/java/com/truelayer/java/entities/HostedPageType.java
@@ -1,0 +1,6 @@
+package com.truelayer.java.entities;
+
+public enum HostedPageType {
+    HPP,
+    HP2
+}