+++ src/main/java/com/truelayer/java/payouts/entities/PayoutUser.java
@@ -1,0 +1,12 @@
+package com.truelayer.java.payouts.entities;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+@EqualsAndHashCode
+public class PayoutUser {
+    private String id;
+}