+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [17.5.0]
+### Added
+* Add support for Verified Payouts with user-determined beneficiary type
+* Add HPP link builder support for payouts
+
 ## [17.4.0] - 2025-08-07
 ### Added
 * Add support for `scheme_id` field in merchant account transaction responses for Payout and Refund transactions